refactor(std/http): move io functions to http/io.ts (#4126)
