+++ std/http/racing_server_test.ts
@@ -1,8 +1,7 @@
-const { connect, run } = Deno;
-
 import { assert, assertEquals } from "../testing/asserts.ts";
 import { BufReader, BufWriter } from "../io/bufio.ts";
 import { TextProtoReader } from "../textproto/mod.ts";
+const { connect, run, test } = Deno;
 
 let server: Deno.Process;
 async function startServer(): Promise<void> {
@@ -59,7 +58,7 @@ content-length: 6
 Step7
 `;
 
-Deno.test(async function serverPipelineRace(): Promise<void> {
+test(async function serverPipelineRace(): Promise<void> {
   await startServer();
 
   const conn = await connect({ port: 4501 });
